+++ src/bioimageio/core/_op_base.py
@@ -66,7 +66,7 @@ class SimpleOperator(BlockwiseOperator):
         elif isinstance(sample, SampleBlock):
             b = sample.blocks[self.input]
             sample.blocks[self.output] = Block(
-                sample_shape=self.get_output_shape(sample.shape[self.input]),
+                sample_shape=self.get_output_shape(sample.sample_shape[self.input]),
                 data=output_tensor,
                 inner_slice=b.inner_slice,
                 halo=b.halo,