+++ src/bioimageio/core/stat_measures.py
@@ -20,7 +20,7 @@ from pydantic import (
     Discriminator,
     PlainSerializer,
 )
-from typing_extensions import Annotated
+from typing_extensions import Annotated, TypeAlias
 
 from .axis import AxisId
 from .common import MemberId, PerMember, QuantileMethod
@@ -40,7 +40,7 @@ def tensor_custom_serializer(t: Tensor) -> Dict[str, Any]:
     return {"data": t.data.data.tolist(), "dims": list(map(str, t.dims))}
 
 
-MeasureValue = Union[
+MeasureValue: TypeAlias = Union[
     float,
     Annotated[
         Tensor,
@@ -180,19 +180,21 @@ class DatasetQuantile(_Quantile, DatasetMeasureBase, frozen=True):
         assert self.axes is None or AxisId("batch") in self.axes
 
 
-SampleMeasure = Annotated[
+SampleMeasure: TypeAlias = Annotated[
     Union[SampleMean, SampleStd, SampleVar, SampleQuantile], Discriminator("name")
 ]
-DatasetMeasure = Annotated[
+DatasetMeasure: TypeAlias = Annotated[
     Union[DatasetMean, DatasetStd, DatasetVar, DatasetQuantile], Discriminator("name")
 ]
-Measure = Annotated[Union[SampleMeasure, DatasetMeasure], Discriminator("scope")]
-Stat = Dict[Measure, MeasureValue]
+Measure: TypeAlias = Annotated[
+    Union[SampleMeasure, DatasetMeasure], Discriminator("scope")
+]
+Stat: TypeAlias = Dict[Measure, MeasureValue]
 
-MeanMeasure = Union[SampleMean, DatasetMean]
-StdMeasure = Union[SampleStd, DatasetStd]
-VarMeasure = Union[SampleVar, DatasetVar]
-PercentileMeasure = Union[SampleQuantile, DatasetQuantile]
+MeanMeasure: TypeAlias = Union[SampleMean, DatasetMean]
+StdMeasure: TypeAlias = Union[SampleStd, DatasetStd]
+VarMeasure: TypeAlias = Union[SampleVar, DatasetVar]
+PercentileMeasure: TypeAlias = Union[SampleQuantile, DatasetQuantile]
 MeanMeasureT = TypeVar("MeanMeasureT", bound=MeanMeasure)
 StdMeasureT = TypeVar("StdMeasureT", bound=StdMeasure)
 VarMeasureT = TypeVar("VarMeasureT", bound=VarMeasure)