+++ changelog.md
@@ -1,3 +1,8 @@
+### 0.10.3
+
+- allow for precomputed sample statistics
+- add `bioimageio.core.io.serialize_stat`
+
 ### 0.10.2
 
 - improve error message when `torch.load(weights_only=True)` fails on non-pure state dicts (#486)